+++ src/UI/NgheJPanel.java
@@ -667,17 +667,20 @@ public class NgheJPanel extends javax.swing.JPanel {
     }
 
     void delete() {
-        if (DialogHelper.confirm(this, "Bạn có muốn xóa hay không?")) {
-            int maba = (int) tblNghe.getValueAt(this.index, 0);
-            try {
-                dao.delete(maba);
-                this.load();
-                clear();
-                DialogHelper.alert(this, "Xóa thành công!");
-            } catch (Exception e) {
-                DialogHelper.alert(this, "Xóa thất bại!");
+        if (ShareHelper.isManager()) {
+            if (DialogHelper.confirm(this, "Bạn có muốn xóa hay không?")) {
+                int maba = (int) tblNghe.getValueAt(this.index, 0);
+                try {
+                    dao.delete(maba);
+                    this.load();
+                    clear();
+                    DialogHelper.alert(this, "Xóa thành công!");
+                } catch (Exception e) {
+                    DialogHelper.alert(this, "Xóa thất bại!");
+                }
             }
+        } else {
+            DialogHelper.alert(this, "Bạn không có quyền xóa");
         }
     }
-
 }
